exclude large block blob test from race detector due to OOM
